manage cancel and send recap mail with sendgrid

+++ src/AzureStorage/StorageService.cs
@@ -8,6 +8,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace AzureStorageAutoBackup.AzureStorage
@@ -30,7 +31,7 @@ namespace AzureStorageAutoBackup.AzureStorage
             _applicationStat = applicationStat;
         }
 
-        public async Task CreateDirectories(List<string> directories)
+        public async Task CreateDirectories(List<string> directories, CancellationTokenSource cancellationToken)
         {
             if (directories.Count == 0)
             {
@@ -46,6 +47,11 @@ namespace AzureStorageAutoBackup.AzureStorage
                 {
                     foreach (var directory in directories.OrderBy(x => x.Length))
                     {
+                        if (cancellationToken.IsCancellationRequested)
+                        {
+                            return;
+                        }
+
                         var path = ToAzureNameWithoutBasePath(directory);
                         var subDirectory = baseDirectory.GetSubdirectoryClient(path);
                         await subDirectory.CreateIfNotExistsAsync();
@@ -54,7 +60,7 @@ namespace AzureStorageAutoBackup.AzureStorage
             }
         }
 
-        public async Task UploadToStorage(FileItem file)
+        public async Task UploadToStorage(FileItem file, CancellationTokenSource cancellationToken)
         {
             var baseDirectory = _shareClient.GetDirectoryClient(_appConfiguration.DestinationPathInAzure);
             var azurePath = ToAzureNameWithoutBasePath(file.Path);
@@ -63,7 +69,14 @@ namespace AzureStorageAutoBackup.AzureStorage
             using (var stream = File.OpenRead(file.Path))
             {
                 fileClient.Create(stream.Length);
-                await UploadFileAsync(fileClient, stream, $"{file.Path} : {file.State}");
+                try
+                {
+                    await UploadFileAsync(cancellationToken, fileClient, stream, $"{file.Path} : {file.State}");
+                }
+                catch (TaskCanceledException)
+                {
+                    return;
+                }
             }
             var metadata = new Dictionary<string, string>
             {
@@ -73,7 +86,7 @@ namespace AzureStorageAutoBackup.AzureStorage
             await _filesState.Save(file);
         }
 
-        private async Task UploadFileAsync(ShareFileClient fileClient, Stream stream, string fileNameAndState)
+        private async Task UploadFileAsync(CancellationTokenSource cancellationToken, ShareFileClient fileClient, Stream stream, string fileNameAndState)
         {
             if (stream.Length == 0)
             {
@@ -92,7 +105,7 @@ namespace AzureStorageAutoBackup.AzureStorage
 
             if (stream.Length <= uploadLimit)
             {
-                await fileClient.UploadRangeAsync(new HttpRange(0, stream.Length), stream, null, progress);
+                await fileClient.UploadRangeAsync(new HttpRange(0, stream.Length), stream, null, progress, cancellationToken: cancellationToken.Token);
                 return;
             }
 
@@ -102,12 +115,12 @@ namespace AzureStorageAutoBackup.AzureStorage
             while ((bytesRead = stream.Read(buffer, 0, buffer.Length)) > 0)
             {
                 using var ms = new MemoryStream(buffer, 0, bytesRead);
-                await fileClient.UploadRangeAsync(ShareFileRangeWriteType.Update, new HttpRange(index, ms.Length), ms, null, progress);
+                await fileClient.UploadRangeAsync(ShareFileRangeWriteType.Update, new HttpRange(index, ms.Length), ms, null, progress, cancellationToken: cancellationToken.Token);
                 index += ms.Length;
             }
         }
 
-        public async Task DeleteFiles(List<string> filesToDelete)
+        public async Task DeleteFiles(List<string> filesToDelete, CancellationTokenSource cancellationTokenSource)
         {
             if (filesToDelete.Count == 0)
             {
@@ -117,6 +130,11 @@ namespace AzureStorageAutoBackup.AzureStorage
             var baseDirectory = _shareClient.GetDirectoryClient(_appConfiguration.DestinationPathInAzure);
             foreach (var file in filesToDelete)
             {
+                if (cancellationTokenSource.IsCancellationRequested)
+                {
+                    return;
+                }
+
                 var path = ToAzureNameWithoutBasePath(file);
                 _applicationStat.DeleteFilesCount++;
                 _logger.LogTrace($"Delete from azure storage {path}");
@@ -125,20 +143,30 @@ namespace AzureStorageAutoBackup.AzureStorage
             }
         }
 
-        public async Task DeleteEmptyDirectoriesIfExist()
+        public async Task DeleteEmptyDirectoriesIfExist(CancellationTokenSource cancellationTokenSource)
         {
             var baseDirectory = _shareClient.GetDirectoryClient(_appConfiguration.DestinationPathInAzure);
-            await DeleteEmptyDirectoriesRecursive(baseDirectory);
+            await DeleteEmptyDirectoriesRecursive(baseDirectory, cancellationTokenSource);
         }
 
-        private async Task DeleteEmptyDirectoriesRecursive(ShareDirectoryClient directoryClient)
+        private async Task DeleteEmptyDirectoriesRecursive(ShareDirectoryClient directoryClient, CancellationTokenSource cancellationTokenSource)
         {
+            if (cancellationTokenSource.IsCancellationRequested)
+            {
+                return;
+            }
+
             await foreach (var item in directoryClient.GetFilesAndDirectoriesAsync())
             {
+                if (cancellationTokenSource.IsCancellationRequested)
+                {
+                    return;
+                }
+
                 if (item.IsDirectory)
                 {
                     var subDirectoryClient = directoryClient.GetSubdirectoryClient(item.Name);
-                    await DeleteEmptyDirectoriesRecursive(subDirectoryClient);
+                    await DeleteEmptyDirectoriesRecursive(subDirectoryClient, cancellationTokenSource);
                 }
             }
 
@@ -161,7 +189,7 @@ namespace AzureStorageAutoBackup.AzureStorage
             }
         }
 
-        public async Task<List<FileItem>> BrowseStorage()
+        public async Task<List<FileItem>> BrowseStorage(CancellationTokenSource cancellationToken)
         {
             var result = new List<FileItem>();
             await _shareClient.CreateIfNotExistsAsync();
@@ -171,20 +199,30 @@ namespace AzureStorageAutoBackup.AzureStorage
                 await baseDirectory.CreateIfNotExistsAsync();
                 if (await baseDirectory.ExistsAsync())
                 {
-                    await RecursiveFindFile(baseDirectory, result);
+                    await RecursiveFindFile(baseDirectory, result, cancellationToken);
                 }
             }
             return result;
         }
 
-        private async Task RecursiveFindFile(ShareDirectoryClient directoryClient, List<FileItem> result)
+        private async Task RecursiveFindFile(ShareDirectoryClient directoryClient, List<FileItem> result, CancellationTokenSource cancellationToken)
         {
+            if (cancellationToken.IsCancellationRequested)
+            {
+                return;
+            }
+
             await foreach (var item in directoryClient.GetFilesAndDirectoriesAsync())
             {
+                if (cancellationToken.IsCancellationRequested)
+                {
+                    return;
+                }
+
                 if (item.IsDirectory)
                 {
                     var subDirectoryClient = directoryClient.GetSubdirectoryClient(item.Name);
-                    await RecursiveFindFile(subDirectoryClient, result);
+                    await RecursiveFindFile(subDirectoryClient, result, cancellationToken);
                 }
                 else
                 {
